Updated Code
This is almost the last bit of code needed for the game to run properly. In this addition I added a function that creates the win screen when a player wins. If the player gets 1000 score the game will end prompting them this screen that tells them they have won. From here they can either restart using the R key or quit using the Q key.